creados los repositorios y cambiadios los id a long

+++ src/main/java/com/argentinaprog/yoprogramo/proyectocv/spa/jere/model/Usuario.java
@@ -19,8 +19,8 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String nombreUsuario;
     private String password;
-    
+
 }
